requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg2-binary>=2.9
openai>=1.0
python-dotenv>=1.0
drf-spectacular>=0.27
social-auth-app-django>=5.4
PyJWT>=2.8
requests>=2.31
